requirements.txt
flask==3.0.0
openpyxl==3.1.2
werkzeug==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.1
